reeply/pyproject.toml
[project]
name = "reeply"
version = "0.1.0"
description = "Manage your texts. Local-first?"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "notion-client>=2.3.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
